Projectionis - Adds this great plugin !
It allows to configure differents projects based on the files in your
working directory.
So we can easly switch between files: source <=> unit_test entity <=>
repository, etc...
It can be reuse by other plugins to find out informations like the
skeleton key.
It was originaly think for the skeleton functionality of Vim but it has
been reused to allow us to load UltiSnips templates when we create a new
file !

A lot of possibilities, really look forward to get used to it !