Completion - NCM2 + phpactor + ncm2-ultisnips
A lot of stuff around completion.
NCM2 is just a completion framework, it doesn't provide any completion
but it improve the completion component a lot.

PhpActor provide completion for PHP.
Along with some refactoring and navigation tools.
It seems promessing but I don't feel like it will be enought just by
itself.
For example, we can't inline an assignation... I used that a lot !
So I might keep php-refactoring-tool under the hood.

ncm2-ultisnips - Great !
It's a wrapper for ultisnips in NCM2, so it provides completion for all
the UltiSnips snippets you have available !
And it provide a completion for function parameters, aloing to jump from
one to another and knowing what you need to provide.
There is another for snipmate, I haven't activated it yet because I
don't know if it can work along UltiSnips without interfering.