Eunuch + Dispatch - Two other tpope plugins
Eunuch is a wrapper around UNIX command, that will allow to do a lot of
stuff a lot more easly.
Ex.: :Move, :Delete
Don't have to :bd the buffer and :!rm <all the path to the file> or
anything like that

Dispatch allows to use vim compilers easly and asynchronously.
It's used by a lot of plugins, I think I first installed it because of a
dependencie :D