EasyMotion - Help when using motion commands by highlightning the possibilies
The plugin specifies it's need to be configurate to be usefull, but I
just want to try and see if there is something for me.
So per default, <Leader><Leader> + motion command to activate it, ex.:
<Leader><Leader>w
<Leader><Leader>f
<Leader><Leader>T

Then it's pretty straight forward.
There is a lot more than that is this plugin, the readme is HUGE, but I
don't have enough time to look into it right know.
Just the motion highlitght could be great for started