EasyAlign - Like Tabular
But different ! :D
Looks like it doesn't need to be configurate, and doesn't need to type a
long command.

You use a mapping to activate it on a given area and then it's
interractive.
Like the idea, will need some time to learn how to use it.
Will try to take it when I have something to align