Mappings - Some new fun stuff
Corrects the mapping ',,' for the number of occurences of the current
search pattern.
Since ',' is my <Leader> it was actually calling EasyMotion.

Adds <Leader>gb and <Leader>gs for, respectivly, :Gblame and :Gstatus

Changes the quick quit insert mapping to 'jk', so I can press them both
at once with one hand only.

The last one, <C-l> to clear the search highlitghing, can't beleive it
took me so long to add it ! :D