plugin - HandleClosingParenthese - Adds a little plugin to handle closing parentheses.
Just a try to see if it's usefull or not.
The goal is to be able to pass threw closing parentheses.
For example, if the cursor is '<>', in:
someFunction(anotherFunction()<>)

I wan't to be able to type ')' and that it just replace the ')' after
the cursor so I can clause the line with ';' without having to leave
insert mode and get to the end of the line.