Add commands for projectionist types using fzf
Projectionist will provide commands like `:E<type>`, this plugin aimes
to do the same with `:<type>`.
Those commands don't need the first letter to indicate how to open the
file because they will only use projectionist to find the list of files
corresponds to a "type" and then provide this list to fzf which will
allow you to open the file in many ways.