Try jiangmiao/auto-pairs to replace kana/vim-smartinput
Smart input works well, maybe once a time I have <C-j> that does not
work as I would like anymore.
So I want to gave a try to this plugin which add some more features,
especially the fly mode that can be nice to exit quicly a block of code.