coc: add autocmd to force signature helper on ,
Because it's my leader and CoC don't want to handle it properly.
Because of the `timeout` and `timeoutlen` `TextChangedI` is not
triggered for every char but only when `timeoutlen` expires
(automatically or because you use a key that does not make a mapping).
But their implementation retrieve the last character in the buffer.

Mean when I type `, `, by the time `TextChangedI` is triggered they
retrieve the space instead of the comma and the signature helper is not
triggered.