fzf: add action to add results to the quickfix
Also configure a big floating window instead of the small one.
I don't really see the point of having a tiny window since the floating
window hide the code beneath it.

I think I prefer to have the usual split window but the default have
change with the new fzf version and I want to gave it a try.
But I don't really like the idea to have my code hidden.
Maybe I will just override the layout for the :Rg command or I will
go back to the split window later on.