ncm2: move after config to the global config file
I had did that so that the autocmd is only declare when ncm2 is
intalled.
At the end I prefer to have all the config in the same file so that I
don't leave any dead config when removing a plugin.

It won't be a problem since when I want to temporarily remove a plugin I
comment all the config file